Create a new project for android logger
