+++ .github/ISSUE_TEMPLATE/bug-report.md
@@ -1,0 +1,28 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+How to reproduce the behavior
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Information (please complete the following information):**
+ - Api Version: [e.g. 21, 22-23, All]
+ - Screen Density: [e.g. mdpi, xhdpi] (Optional)
+ - Device Model: [e.g. One Plus 5T] (Optional)
+
+**Additional context**
+Add any other context about the problem here.
