+++ .github/ISSUE_TEMPLATE/documentation-request.md
@@ -1,0 +1,12 @@
+---
+name: Documentation request
+about: Help us improving our docs and examples
+title: ''
+labels: documentation
+assignees: ''
+
+---
+
+**Is something in our documentation missing or not well explained? Do you need help?**
+
+Please describe how can we improve our documentation
