+++ CONTRIBUTING.md
@@ -1,0 +1,26 @@
+# Contributing to mistica-android
+
+Thank you for your interest in this repository. You can contribute to this project in several ways:
+
+## Pull Requests
+
+We would love to accept your Pull Requests but please, before starting your development,
+[create an issue](https://github.com/Telefonica/android-logger/issues/new/choose).
+
+## Bug reports
+
+If something is broken or not working as expected, let us know!
+
+:bug: [Open a Bug issue](https://github.com/Telefonica/android-logger/issues/new?assignees=&labels=bug&template=bug-report.md&title=)
+
+## Feature requests (no UI/UX changes)
+
+If you need additional functionality, support a new use case...
+
+:construction: [Open a Feature Request issue](https://github.com/Telefonica/android-logger/issues/new?assignees=&labels=enhancement&template=feature-request.md&title=)
+
+## Documentation and help requests
+
+Is something in our documentation not well explained? Do you need help? 
+
+:blue_book: [Open an Documentation issue](https://github.com/Telefonica/android-logger/issues/new?assignees=&labels=documentation&template=documentation-request.md&title=)
