+++ library/src/enabled/java/com/telefonica/androidlogger/io/provider/AppLoggerFileProvider.kt
@@ -1,0 +1,5 @@
+package com.telefonica.androidlogger.io.provider
+
+import androidx.core.content.FileProvider
+
+class AppLoggerFileProvider : FileProvider()
