--- library/src/enabled/java/com/tuenti/applogger/io/provider/AppLoggerFileProvider.kt
@@ -1,5 +1,0 @@
-package com.tuenti.applogger.io.provider
-
-import androidx.core.content.FileProvider
-
-class AppLoggerFileProvider : FileProvider()
