+++ library/src/main/java/com/telefonica/androidlogger/io/executor/TaskCallback.kt
@@ -1,0 +1,12 @@
+package com.telefonica.androidlogger.io.executor
+
+class TaskCallback<T>(
+    func: (T?) -> Unit
+) {
+    @Volatile
+    internal var func: ((T?) -> Unit)? = func
+
+    fun cancel() {
+        func = null
+    }
+}
