--- library/src/main/java/com/tuenti/applogger/io/executor/TaskCallback.kt
@@ -1,12 +1,0 @@
-package com.tuenti.applogger.io.executor
-
-class TaskCallback<T>(
-    func: (T?) -> Unit
-) {
-    @Volatile
-    internal var func: ((T?) -> Unit)? = func
-
-    fun cancel() {
-        func = null
-    }
-}
