+++ library/src/noop/java/com/telefonica/androidlogger/ui/AppLoggerActivity.kt
@@ -1,0 +1,9 @@
+package com.telefonica.androidlogger.ui
+
+import android.content.Context
+import android.content.Intent
+import com.telefonica.androidlogger.domain.LogCategory
+
+@Suppress("UNUSED_PARAMETER")
+fun getLaunchIntent(context: Context, categories: List<LogCategory>): Intent? =
+    null
