--- library/src/noop/java/com/tuenti/applogger/ui/AppLoggerActivity.kt
@@ -1,9 +1,0 @@
-package com.tuenti.applogger.ui
-
-import android.content.Context
-import android.content.Intent
-import com.tuenti.applogger.domain.LogCategory
-
-@Suppress("UNUSED_PARAMETER")
-fun getLaunchIntent(context: Context, categories: List<LogCategory>): Intent? =
-    null
